change xbr name spacing
